requirements.txt
python-dotenv
streamlit
langchain
langchain-openai
langchain-community
beautifulsoup4
chromadb
llama_index
pymupdf
unstructured
openpyxl
markdown
sentence_transformers